refactor: passar header buttons como children p/ GameHeader
