+++ src/components/Game/components/CheckGameButton/index.tsx
