+++ src/components/Game/components/CheckGameButton/style.css
