+++ src/components/Game/components/CheckGameButton/style.ts
