+++ src/components/Game/components/SelectNewGameButton/style.ts
