+++ src/components/Game/components/index.ts
