+++ src/components/Game/index.tsx
@@ -5,6 +5,7 @@ import { GameBoard } from '../GameBoard';
 import { Timer } from '../Timer';
 import { useTimer } from '../Timer/hooks';
 import { gameDisplayStyle, gameStyle } from './style';
+import { CheckGameButton, SelectNewGameButton } from './components';
 import './style.css';
 
 export const Game: FC = () => {
@@ -16,6 +17,8 @@ export const Game: FC = () => {
         <div className="game" style={gameStyle}>
             <GameHeader>
                 <Timer {...{ counter, startStopTimer, resetTimer, isOn }} />
+                <CheckGameButton />
+                <SelectNewGameButton />
             </GameHeader>
             <div id="game-display" style={gameDisplayStyle}>
                 <GameBoard
