+++ src/components/GameHeader/index.tsx
@@ -1,6 +1,5 @@
 import React, { FC, ReactNode } from 'react';
-import { CheckGameButton, SelectNewGameButton } from './components';
-import { gameHeaderStyle, gameInfoStyle, gamePageStyle } from './style';
+import { gameHeaderStyle, gamePageStyle } from './style';
 
 export type GameHeaderProps = {
     children: ReactNode;
@@ -10,11 +9,7 @@ export const GameHeader: FC<GameHeaderProps> = ({ children }) => {
     return (
         <div id="game-page" style={gamePageStyle}>
             <div id="game-header" style={gameHeaderStyle}>
-                <div id="game-info" style={gameInfoStyle}>
-                    {children}
-                </div>
-                <CheckGameButton />
-                <SelectNewGameButton />
+                {children}
             </div>
         </div>
     );
