+++ src/components/Timer/index.tsx
@@ -1,5 +1,5 @@
 import React, { FC } from 'react';
-import { timerButtonStyle, timerStyle } from './style';
+import { gameInfoStyle, timerButtonStyle, timerStyle } from './style';
 import { format } from 'date-fns';
 import './style.css';
 import { useTimer } from './hooks';
@@ -13,27 +13,29 @@ export const Timer: FC<TimerProps> = ({
     isOn,
 }) => {
     return (
-        <div id="timer" style={timerStyle}>
-            <p style={{ fontWeight: 'bolder' }}>
-                {counter === 0
-                    ? format(new Date().setHours(0, 0, 0), 'HH:mm:ss')
-                    : format(counter, 'HH:mm:ss')}
-            </p>
-            <button
-                id="start-stop-btn"
-                className="stop"
-                style={timerButtonStyle}
-                onClick={startStopTimer}
-            >
-                {isOn ? 'Stop' : 'Start'}
-            </button>
-            <button
-                id="reset-btn"
-                style={timerButtonStyle}
-                onClick={resetTimer}
-            >
-                Reset
-            </button>
+        <div id="game-info" style={gameInfoStyle}>
+            <div id="timer" style={timerStyle}>
+                <p style={{ fontWeight: 'bolder' }}>
+                    {counter === 0
+                        ? format(new Date().setHours(0, 0, 0), 'HH:mm:ss')
+                        : format(counter, 'HH:mm:ss')}
+                </p>
+                <button
+                    id="start-stop-btn"
+                    className="stop"
+                    style={timerButtonStyle}
+                    onClick={startStopTimer}
+                >
+                    {isOn ? 'Stop' : 'Start'}
+                </button>
+                <button
+                    id="reset-btn"
+                    style={timerButtonStyle}
+                    onClick={resetTimer}
+                >
+                    Reset
+                </button>
+            </div>
         </div>
     );
 };
