+++ src/components/Timer/style.ts
@@ -1,5 +1,14 @@
 import { CSSProperties } from 'react';
 
+export const gameInfoStyle: CSSProperties = {
+    border: '1px solid black',
+    borderRadius: '5px',
+    display: 'flex',
+    flexDirection: 'column',
+    padding: '0.2rem',
+    marginBottom: '1px',
+};
+
 export const timerStyle: CSSProperties = {
     display: 'flex',
     flexDirection: 'row',
